Do not clear the cache !
Currently, clearing the shared cache yield to situation vers matching functions are not generated.
(when a module starts being compiled, it clear the cache, possibly leaving other module compilations into a unspecified state).
A side effect of not clearinig the cache (aside avoiding not being able to compile) is that each module will have
functions for ALL entries of the cache. Which is not desirable..
One way to solved this this problem is to to create a per process cache system.